chore: Extract scope manager stack
